backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
pydantic==2.8.2
pydantic-settings==2.4.0
SQLAlchemy==2.0.36
psycopg[binary,pool]==3.2.1
alembic==1.13.2
passlib[bcrypt]==1.7.4
argon2-cffi==23.1.0
python-jose[cryptography]==3.3.0
email-validator==2.2.0
python-multipart
weasyprint
jinja2
httpx>=0.27,<0.28